circuits: another bounds helper and doc comment on validate
